// Frontend/src/app/models/useradd.model.ts
export class UserAdd {
    userId?: number = 0;
    firstname?: string = "";
    lastname?: string = "";
    email?: string = "";
    username?: string = "";
    password?: string = "";
    confirmPassword?: string = "";
    isActive?: boolean = true;
    roleId?: number = 0;
    //roleType?: string = "";
}